Improve dedupping probability on load+store (#31)
Use case: we load from some non-recent slot "EARLY", do something which
doesn't change the files, and then store to the slot "LATE".

Previously, when rsyncing-out to slot "LATE", we tried to hardlink-dedup
with the **most recent** slot on ci-storage end (let's name it
"MIDDLE"). The problem is that, since we actually loaded not from the
most recent "MIDDLE" slot, but from another ("EARLY") slot, then, when
storing, we didn't get optimal hardlink-dedupping, because unchanged
files in "LATE" are technically compared with "MIDDLE" and not with
"EARLY" data, where they did originate from.

What's funny is that we store the real loading history in the slot meta,
but we did not use it here.

In this PR, the history info starts being used to detect the best slot
to dedup with ("EARLY" in the above example). I.e. if we loaded from
somewhere, then we try to dedup with that exact "somewhere" when
storing.

## PRs in the Stack
- ➡ #31

(The stack is managed by
[git-grok](https://github.com/dimikot/git-grok).)